jointiur/requirements.txt
hydra-core==1.1.1
hydra-core==1.1.1
transformers==4.11.3
torch==1.9.1
nltk==3.6.1
fasttext==0.9.2
mlflow==1.20.2
stopwordsiso==0.6.1
jieba==0.42.1
mlflow==1.20.2
rouge==1.0.1
SentencePiece==0.1.96
omegaconf==2.1.1